Minor changes
Added commentary
Changed filelength type from int to long

+++ FileMetaData.java
@@ -1,11 +1,12 @@
 package BACKUPER;
 
 public class FileMetaData {
-    private int fileLength;
+    private long fileLength;
     private String filePath;
     private String clientPath;
     private int version;
-    public int getFileLength() {
+    
+    public long getFileLength() {
         return this.fileLength;
     }
     public String getFilePath() {
@@ -19,13 +20,13 @@ public class FileMetaData {
         return version;
     }
 
-    public FileMetaData(String filePath, int fileLength){
+    public FileMetaData(String filePath, long fileLength){
         this.fileLength = fileLength;
         this.filePath = filePath;
         this.clientPath="";
         this.version=0;
     }
-    public FileMetaData(int version,String clientPath,String filePath,int fileLength){
+    public FileMetaData(int version,String clientPath,String filePath,long fileLength){
         this.fileLength=fileLength;
         this.filePath=filePath;
         this.clientPath=clientPath;
